+++ src/main/java/org/asamk/signal/Main.java
@@ -20,20 +20,13 @@ import net.sourceforge.argparse4j.ArgumentParsers;
 import net.sourceforge.argparse4j.impl.Arguments;
 import net.sourceforge.argparse4j.inf.ArgumentParser;
 import net.sourceforge.argparse4j.inf.ArgumentParserException;
-import net.sourceforge.argparse4j.inf.MutuallyExclusiveGroup;
 import net.sourceforge.argparse4j.inf.Namespace;
-import net.sourceforge.argparse4j.inf.Subparser;
-import net.sourceforge.argparse4j.inf.Subparsers;
 
-import org.asamk.signal.commands.Command;
-import org.asamk.signal.commands.Commands;
 import org.asamk.signal.manager.LibSignalLogger;
 import org.asamk.signal.util.SecurityProvider;
 import org.bouncycastle.jce.provider.BouncyCastleProvider;
-import org.whispersystems.signalservice.api.util.PhoneNumberFormatter;
 
 import java.security.Security;
-import java.util.Map;
 
 public class Main {
 
@@ -41,26 +34,13 @@ public class Main {
         installSecurityProviderWorkaround();
 
         // Configuring the logger needs to happen before any logger is initialized
-        if (isVerbose(args)) {
-            System.setProperty("org.slf4j.simpleLogger.defaultLogLevel", "trace");
-            System.setProperty("org.slf4j.simpleLogger.showThreadName", "true");
-            System.setProperty("org.slf4j.simpleLogger.showShortLogName", "false");
-            System.setProperty("org.slf4j.simpleLogger.showDateTime", "true");
-            System.setProperty("org.slf4j.simpleLogger.dateTimeFormat", "yyyy-MM-dd'T'HH:mm:ss.SSSXX");
-            LibSignalLogger.initLogger();
-        } else {
-            System.setProperty("org.slf4j.simpleLogger.defaultLogLevel", "info");
-            System.setProperty("org.slf4j.simpleLogger.showThreadName", "false");
-            System.setProperty("org.slf4j.simpleLogger.showShortLogName", "true");
-            System.setProperty("org.slf4j.simpleLogger.showDateTime", "false");
-        }
+        configureLogging(isVerbose(args));
 
-        Namespace ns = parseArgs(args);
-        if (ns == null) {
-            System.exit(1);
-        }
+        ArgumentParser parser = App.buildArgumentParser();
 
-        int res = new Cli(ns).init();
+        Namespace ns = parser.parseArgsOrFail(args);
+
+        int res = new App(ns).init();
         System.exit(res);
     }
 
@@ -71,7 +51,8 @@ public class Main {
     }
 
     private static boolean isVerbose(String[] args) {
-        ArgumentParser parser = buildBaseArgumentParser();
+        ArgumentParser parser = ArgumentParsers.newFor("signal-cli").build().defaultHelp(false);
+        parser.addArgument("--verbose").action(Arguments.storeTrue());
 
         Namespace ns;
         try {
@@ -83,87 +64,19 @@ public class Main {
         return ns.getBoolean("verbose");
     }
 
-    private static Namespace parseArgs(String[] args) {
-        ArgumentParser parser = buildArgumentParser();
-
-        Namespace ns;
-        try {
-            ns = parser.parseArgs(args);
-        } catch (ArgumentParserException e) {
-            parser.handleError(e);
-            return null;
+    private static void configureLogging(final boolean verbose) {
+        if (verbose) {
+            System.setProperty("org.slf4j.simpleLogger.defaultLogLevel", "debug");
+            System.setProperty("org.slf4j.simpleLogger.showThreadName", "true");
+            System.setProperty("org.slf4j.simpleLogger.showShortLogName", "false");
+            System.setProperty("org.slf4j.simpleLogger.showDateTime", "true");
+            System.setProperty("org.slf4j.simpleLogger.dateTimeFormat", "yyyy-MM-dd'T'HH:mm:ss.SSSXX");
+            LibSignalLogger.initLogger();
+        } else {
+            System.setProperty("org.slf4j.simpleLogger.defaultLogLevel", "info");
+            System.setProperty("org.slf4j.simpleLogger.showThreadName", "false");
+            System.setProperty("org.slf4j.simpleLogger.showShortLogName", "true");
+            System.setProperty("org.slf4j.simpleLogger.showDateTime", "false");
         }
-
-        if ("link".equals(ns.getString("command"))) {
-            if (ns.getString("username") != null) {
-                parser.printUsage();
-                System.err.println("You cannot specify a username (phone number) when linking");
-                System.exit(2);
-            }
-        } else if (!ns.getBoolean("dbus") && !ns.getBoolean("dbus_system")) {
-            if (ns.getString("username") == null) {
-                parser.printUsage();
-                System.err.println("You need to specify a username (phone number)");
-                System.exit(2);
-            }
-            if (!PhoneNumberFormatter.isValidNumber(ns.getString("username"), null)) {
-                System.err.println("Invalid username (phone number), make sure you include the country code.");
-                System.exit(2);
-            }
-        }
-
-        if (ns.getList("recipient") != null && !ns.getList("recipient").isEmpty() && ns.getString("group") != null) {
-            System.err.println("You cannot specify recipients by phone number and groups at the same time");
-            System.exit(2);
-        }
-
-        return ns;
-    }
-
-    private static ArgumentParser buildArgumentParser() {
-        ArgumentParser parser = buildBaseArgumentParser();
-
-        Subparsers subparsers = parser.addSubparsers()
-                .title("subcommands")
-                .dest("command")
-                .description("valid subcommands")
-                .help("additional help");
-
-        final Map<String, Command> commands = Commands.getCommands();
-        for (Map.Entry<String, Command> entry : commands.entrySet()) {
-            Subparser subparser = subparsers.addParser(entry.getKey());
-            entry.getValue().attachToSubparser(subparser);
-        }
-
-        return parser;
-    }
-
-    private static ArgumentParser buildBaseArgumentParser() {
-        ArgumentParser parser = ArgumentParsers.newFor("signal-cli")
-                .build()
-                .defaultHelp(true)
-                .description("Commandline interface for Signal.")
-                .version(BaseConfig.PROJECT_NAME + " " + BaseConfig.PROJECT_VERSION);
-
-        parser.addArgument("-v", "--version").help("Show package version.").action(Arguments.version());
-        parser.addArgument("--verbose")
-                .help("Raise log level and include lib signal logs.")
-                .action(Arguments.storeTrue());
-        parser.addArgument("--config")
-                .help("Set the path, where to store the config (Default: $XDG_DATA_HOME/signal-cli , $HOME/.local/share/signal-cli).");
-        parser.addArgument("-n", "--busname")
-                .help("Name of the DBus.");
-
-        MutuallyExclusiveGroup mut = parser.addMutuallyExclusiveGroup();
-        mut.addArgument("-u", "--username").help("Specify your phone number, that will be used for verification.");
-        mut.addArgument("--dbus").help("Make request via user dbus.").action(Arguments.storeTrue());
-        mut.addArgument("--dbus-system").help("Make request via system dbus.").action(Arguments.storeTrue());
-
-        parser.addArgument("-o", "--output")
-                .help("Choose to output in plain text or JSON")
-                .choices("plain-text", "json")
-                .setDefault("plain-text");
-
-        return parser;
     }
 }