+++ src/main/java/org/asamk/signal/commands/DaemonCommand.java
@@ -4,8 +4,10 @@ import net.sourceforge.argparse4j.impl.Arguments;
 import net.sourceforge.argparse4j.inf.Namespace;
 import net.sourceforge.argparse4j.inf.Subparser;
 
+import org.asamk.signal.DbusConfig;
 import org.asamk.signal.DbusReceiveMessageHandler;
 import org.asamk.signal.JsonDbusReceiveMessageHandler;
+import org.asamk.signal.OutputType;
 import org.asamk.signal.dbus.DbusSignalImpl;
 import org.asamk.signal.manager.Manager;
 import org.freedesktop.dbus.connections.impl.DBusConnection;
@@ -14,15 +16,14 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Set;
 import java.util.concurrent.TimeUnit;
 
-import static org.asamk.signal.DbusConfig.SIGNAL_BUSNAME;
-import static org.asamk.signal.DbusConfig.SIGNAL_OBJECTPATH;
-import static org.asamk.signal.util.ErrorUtils.handleAssertionError;
+public class DaemonCommand implements MultiLocalCommand {
 
-public class DaemonCommand implements LocalCommand {
-
-    private final static Logger logger = LoggerFactory.getLogger(ReceiveCommand.class);
+    private final static Logger logger = LoggerFactory.getLogger(DaemonCommand.class);
 
     @Override
     public void attachToSubparser(final Subparser subparser) {
@@ -38,60 +39,123 @@ public class DaemonCommand implements LocalCommand {
     }
 
     @Override
+    public Set<OutputType> getSupportedOutputTypes() {
+        return Set.of(OutputType.PLAIN_TEXT, OutputType.JSON);
+    }
+
+    @Override
     public int handleCommand(final Namespace ns, final Manager m) {
-        boolean inJson = ns.getString("output").equals("json") || ns.getBoolean("json");
+        boolean inJson = ns.get("output") == OutputType.JSON || ns.getBoolean("json");
 
         // TODO delete later when "json" variable is removed
         if (ns.getBoolean("json")) {
             logger.warn("\"--json\" option has been deprecated, please use the global \"--output=json\" instead.");
         }
 
-        DBusConnection conn = null;
-        try {
-            try {
-                DBusConnection.DBusBusType busType;
-                String  busName;
-                if (ns.getBoolean("system")) {
-                    busType = DBusConnection.DBusBusType.SYSTEM;
-                } else {
-                    busType = DBusConnection.DBusBusType.SESSION;
-                }
-                conn = DBusConnection.getConnection(busType);
-                conn.exportObject(SIGNAL_OBJECTPATH, new DbusSignalImpl(m));
-		busName = ns.getString("busname");
-		if (busName == null) {
-		    conn.requestBusName(SIGNAL_BUSNAME);
-		} else {
-		    conn.requestBusName(busName);
-		}
-            } catch (UnsatisfiedLinkError e) {
-                System.err.println("Missing native library dependency for dbus service: " + e.getMessage());
-                return 1;
-            } catch (DBusException e) {
-                e.printStackTrace();
-                return 2;
-            }
-            boolean ignoreAttachments = ns.getBoolean("ignore_attachments");
-            try {
-                m.receiveMessages(1,
-                        TimeUnit.HOURS,
-                        false,
-                        ignoreAttachments,
-                        inJson
-                                ? new JsonDbusReceiveMessageHandler(m, conn, SIGNAL_OBJECTPATH)
-                                : new DbusReceiveMessageHandler(m, conn, SIGNAL_OBJECTPATH));
-                return 0;
-            } catch (IOException e) {
-                System.err.println("Error while receiving messages: " + e.getMessage());
-                return 3;
-            } catch (AssertionError e) {
-                handleAssertionError(e);
-                return 1;
-            }
-        } finally {
-            if (conn != null) {
-                conn.disconnect();
-            }
+        boolean ignoreAttachments = ns.getBoolean("ignore_attachments");
+
+        DBusConnection.DBusBusType busType;
+        if (ns.getBoolean("system")) {
+            busType = DBusConnection.DBusBusType.SYSTEM;
+        } else {
+            busType = DBusConnection.DBusBusType.SESSION;
         }
+
+        try (DBusConnection conn = DBusConnection.getConnection(busType)) {
+            String objectPath = DbusConfig.getObjectPath();
+            String  busName;
+            Thread t = run(conn, objectPath, m, ignoreAttachments, inJson);
+
+            busName = ns.getString("busname");
+            if (busName == null) {
+                conn.requestBusName(DbusConfig.getBusname());
+            } else {
+                conn.requestBusName(busName);
+            }
+
+            try {
+                t.join();
+            } catch (InterruptedException ignored) {
+            }
+            return 0;
+        } catch (DBusException | IOException e) {
+            logger.error("Dbus command failed", e);
+            return 2;
+        }
+    }
+
+    @Override
+    public int handleCommand(final Namespace ns, final List<Manager> managers) {
+        boolean inJson = ns.get("output") == OutputType.JSON || ns.getBoolean("json");
+
+        // TODO delete later when "json" variable is removed
+        if (ns.getBoolean("json")) {
+            logger.warn("\"--json\" option has been deprecated, please use the global \"--output=json\" instead.");
+        }
+
+        boolean ignoreAttachments = ns.getBoolean("ignore_attachments");
+
+        DBusConnection.DBusBusType busType;
+        if (ns.getBoolean("system")) {
+            busType = DBusConnection.DBusBusType.SYSTEM;
+        } else {
+            busType = DBusConnection.DBusBusType.SESSION;
+        }
+
+        try (DBusConnection conn = DBusConnection.getConnection(busType)) {
+            List<Thread> receiveThreads = new ArrayList<>();
+            String  busName;
+            for (Manager m : managers) {
+                String objectPath = DbusConfig.getObjectPath(m.getUsername());
+                Thread thread = run(conn, objectPath, m, ignoreAttachments, inJson);
+                receiveThreads.add(thread);
+            }
+
+            busName = ns.getString("busname");
+            if (busName == null) {
+                conn.requestBusName(DbusConfig.getBusname());
+            } else {
+                conn.requestBusName(busName);
+            }
+
+            for (Thread t : receiveThreads) {
+                try {
+                    t.join();
+                } catch (InterruptedException ignored) {
+                }
+            }
+            return 0;
+        } catch (DBusException | IOException e) {
+            logger.error("Dbus command failed", e);
+            return 2;
+        }
+    }
+
+    private Thread run(
+            DBusConnection conn, String objectPath, Manager m, boolean ignoreAttachments, boolean inJson
+    ) throws DBusException {
+        conn.exportObject(objectPath, new DbusSignalImpl(m));
+
+        final Thread thread = new Thread(() -> {
+            while (true) {
+                try {
+                    m.receiveMessages(1,
+                            TimeUnit.HOURS,
+                            false,
+                            ignoreAttachments,
+                            inJson
+                                    ? new JsonDbusReceiveMessageHandler(m, conn, objectPath)
+                                    : new DbusReceiveMessageHandler(m, conn, objectPath));
+                } catch (IOException e) {
+                    logger.warn("Receiving messages failed, retrying", e);
+                }
+            }
+        });
+
+        logger.info("Exported dbus object: " + objectPath);
+
+        thread.start();
+
+        return thread;
     }
 }