+++ src/main/java/org/asamk/signal/commands/ListGroupsCommand.java
@@ -4,15 +4,14 @@ import net.sourceforge.argparse4j.impl.Arguments;
 import net.sourceforge.argparse4j.inf.Namespace;
 import net.sourceforge.argparse4j.inf.Subparser;
 
+import org.asamk.signal.JsonWriter;
+import org.asamk.signal.OutputType;
 import org.asamk.signal.manager.Manager;
 import org.asamk.signal.manager.groups.GroupInviteLinkUrl;
 import org.asamk.signal.manager.storage.groups.GroupInfo;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.whispersystems.signalservice.api.push.SignalServiceAddress;
-
-import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import com.fasterxml.jackson.annotation.PropertyAccessor;
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -22,22 +21,13 @@ import java.util.stream.Collectors;
 
 public class ListGroupsCommand implements LocalCommand {
 
+    private final static Logger logger = LoggerFactory.getLogger(ListGroupsCommand.class);
+
     private static Set<String> resolveMembers(Manager m, Set<SignalServiceAddress> addresses) {
-        return addresses.stream().map(m::resolveSignalServiceAddress)
+        return addresses.stream()
+                .map(m::resolveSignalServiceAddress)
                 .map(SignalServiceAddress::getLegacyIdentifier)
                 .collect(Collectors.toSet());
-    }
-
-    private static int printGroupsJson(ObjectMapper jsonProcessor, List<?> objects) {
-        try {
-            jsonProcessor.writeValue(System.out, objects);
-            System.out.println();
-        } catch (IOException e) {
-            System.err.println(e.getMessage());
-            return 1;
-        }
-
-        return 0;
     }
 
     private static void printGroupPlainText(Manager m, GroupInfo group, boolean detailed) {
@@ -65,24 +55,28 @@ public class ListGroupsCommand implements LocalCommand {
 
     @Override
     public void attachToSubparser(final Subparser subparser) {
-        subparser.addArgument("-d", "--detailed").action(Arguments.storeTrue())
+        subparser.addArgument("-d", "--detailed")
+                .action(Arguments.storeTrue())
                 .help("List the members and group invite links of each group. If output=json, then this is always set");
 
         subparser.help("List group information including names, ids, active status, blocked status and members");
     }
 
     @Override
-    public int handleCommand(final Namespace ns, final Manager m) {
-        if (ns.getString("output").equals("json")) {
-            final ObjectMapper jsonProcessor = new ObjectMapper();
-            jsonProcessor.setVisibility(PropertyAccessor.ALL, JsonAutoDetect.Visibility.ANY);
-            jsonProcessor.disable(JsonGenerator.Feature.AUTO_CLOSE_TARGET);
+    public Set<OutputType> getSupportedOutputTypes() {
+        return Set.of(OutputType.PLAIN_TEXT, OutputType.JSON);
+    }
 
-            List<JsonGroup> objects = new ArrayList<>();
+    @Override
+    public int handleCommand(final Namespace ns, final Manager m) {
+        if (ns.get("output") == OutputType.JSON) {
+            final JsonWriter jsonWriter = new JsonWriter(System.out);
+
+            List<JsonGroup> jsonGroups = new ArrayList<>();
             for (GroupInfo group : m.getGroups()) {
                 final GroupInviteLinkUrl groupInviteLink = group.getGroupInviteLink();
 
-                objects.add(new JsonGroup(group.getGroupId().toBase64(),
+                jsonGroups.add(new JsonGroup(group.getGroupId().toBase64(),
                         group.getTitle(),
                         group.isMember(m.getSelfAddress()),
                         group.isBlocked(),
@@ -91,7 +85,15 @@ public class ListGroupsCommand implements LocalCommand {
                         resolveMembers(m, group.getRequestingMembers()),
                         groupInviteLink == null ? null : groupInviteLink.getUrl()));
             }
-            return printGroupsJson(jsonProcessor, objects);
+
+            try {
+                jsonWriter.write(jsonGroups);
+            } catch (IOException e) {
+                logger.error("Failed to write json object: {}", e.getMessage());
+                return 3;
+            }
+
+            return 0;
         } else {
             boolean detailed = ns.getBoolean("detailed");
             for (GroupInfo group : m.getGroups()) {
@@ -114,10 +116,16 @@ public class ListGroupsCommand implements LocalCommand {
         public Set<String> requestingMembers;
         public String groupInviteLink;
 
-        public JsonGroup(String id, String name, boolean isMember, boolean isBlocked,
-                         Set<String> members, Set<String> pendingMembers,
-                         Set<String> requestingMembers, String groupInviteLink)
-        {
+        public JsonGroup(
+                String id,
+                String name,
+                boolean isMember,
+                boolean isBlocked,
+                Set<String> members,
+                Set<String> pendingMembers,
+                Set<String> requestingMembers,
+                String groupInviteLink
+        ) {
             this.id = id;
             this.name = name;
             this.isMember = isMember;