+++ src/main/java/org/asamk/signal/dbus/DbusSignalImpl.java
@@ -45,9 +45,30 @@ public class DbusSignalImpl implements Signal {
         return sendMessage(message, attachments, recipients);
     }
 
+    private static void checkSendMessageResult(long timestamp, SendMessageResult result) throws DBusExecutionException {
+        String error = ErrorUtils.getErrorMessageFromSendMessageResult(result);
+
+        if (error == null) {
+            return;
+        }
+
+        final String message = timestamp + "\nFailed to send message:\n" + error + '\n';
+
+        if (result.getIdentityFailure() != null) {
+            throw new Error.UntrustedIdentity(message);
+        } else {
+            throw new Error.Failure(message);
+        }
+    }
+
     private static void checkSendMessageResults(
             long timestamp, List<SendMessageResult> results
     ) throws DBusExecutionException {
+        if (results.size() == 1) {
+            checkSendMessageResult(timestamp, results.get(0));
+            return;
+        }
+
         List<String> errors = ErrorUtils.getErrorMessagesFromSendMessageResults(results);
         if (errors.size() == 0) {
             return;
@@ -71,6 +92,21 @@ public class DbusSignalImpl implements Signal {
             return results.first();
         } catch (InvalidNumberException e) {
             throw new Error.InvalidNumber(e.getMessage());
+        } catch (AttachmentInvalidException e) {
+            throw new Error.AttachmentInvalid(e.getMessage());
+        } catch (IOException e) {
+            throw new Error.Failure(e.getMessage());
+        }
+    }
+
+    @Override
+    public long sendNoteToSelfMessage(
+            final String message, final List<String> attachments
+    ) throws Error.AttachmentInvalid, Error.Failure, Error.UntrustedIdentity {
+        try {
+            final Pair<Long, SendMessageResult> results = m.sendSelfMessage(message, attachments);
+            checkSendMessageResult(results.first(), results.second());
+            return results.first();
         } catch (AttachmentInvalidException e) {
             throw new Error.AttachmentInvalid(e.getMessage());
         } catch (IOException e) {